Eventos de Dominio implmentados para el proceso de actualización del estado de una Configuración después de una operación crítica

+++ src/configs/herramientas/herramientas.service.ts
@@ -1,11 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Herramienta, TipoHerramienta } from './herramienta.entity';
 import { Repository } from 'typeorm';
+import { EventEmitter2 } from '@nestjs/event-emitter';
 
 @Injectable()
 export class HerramientasService {
-    constructor(@InjectRepository(Herramienta) private herramientaRepository: Repository<Herramienta>) { }
+    constructor(@InjectRepository(Herramienta) private herramientaRepository: Repository<Herramienta>, private eventEmitter: EventEmitter2) { }
 
 
     // Método para obtener una herramienta
@@ -29,6 +30,19 @@ export class HerramientasService {
 
     // Método para eliminar una herramienta en específico
     public async deleteHerramienta(idHerramienta: number) {
-        await this.herramientaRepository.delete({ id: idHerramienta })
+         // se obtiene primero la herramienta  antes de eliminarlo
+         const herramienta: Herramienta | undefined = await this.herramientaRepository.findOne({
+            where: {
+                id: idHerramienta
+            }
+        })
+
+        if (herramienta) {
+            await this.herramientaRepository.delete({ id: idHerramienta })
+            // se emite el evento
+            await this.eventEmitter.emitAsync("accionCritica", herramienta.configVersion)
+        }
+        else
+            throw new HttpException("No existe una Herramienta con ese id", HttpStatus.BAD_REQUEST)
     }
 }
